ObjectComparison task
