+++ cs/HomeExercisesTests/ObjectComparison.cs
@@ -1,0 +1,63 @@
+using FluentAssertions;
+using HomeExercises;
+using NUnit.Framework;
+
+namespace HomeExercisesTests
+{
+    public class ObjectComparison
+    {
+        [Test]
+        [Description("Проверка текущего царя")]
+        [Category("ToRefactor")]
+        public void CheckCurrentTsar()
+        {
+            var actualTsar = TsarRegistry.GetCurrentTsar();
+
+            var expectedTsar = new Person("Ivan IV The Terrible", 54, 170, 70,
+                new Person("Vasili III of Russia", 28, 170, 60, null));
+
+            // Перепишите код на использование Fluent Assertions.
+            actualTsar.Should().BeEquivalentTo(expectedTsar, config
+                => config.Excluding(member
+                => member.SelectedMemberInfo.Name == nameof(expectedTsar.Id)));
+        }
+
+        [Test]
+        [Description("Альтернативное решение. Какие у него недостатки?")]
+        public void CheckCurrentTsar_WithCustomEquality()
+        {
+            var actualTsar = TsarRegistry.GetCurrentTsar();
+            var expectedTsar = new Person("Ivan IV The Terrible", 54, 170, 70,
+                new Person("Vasili III of Russia", 28, 170, 60, null));
+
+            // Какие недостатки у такого подхода? 
+            //Сравнения полей захардкожены, при изменении Person метод придется переписывать
+            //Возможная рекурсия при сравнении потомков
+            //Тяжело дебажить, непонятно, где будет падать тест
+            //Ну и метод AreEquals, работающий с Person выглядит странно в классе с тестами
+            Assert.True(AreEqual(actualTsar, expectedTsar));
+        }
+
+        private bool AreEqual(Person? actual, Person? expected)
+        {
+            if (actual == expected) return true;
+            if (actual == null || expected == null) return false;
+            return
+                actual.Name == expected.Name
+                && actual.Age == expected.Age
+                && actual.Height == expected.Height
+                && actual.Weight == expected.Weight
+                && AreEqual(actual.Parent, expected.Parent);
+        }
+    }
+
+    public class TsarRegistry
+    {
+        public static Person GetCurrentTsar()
+        {
+            return new Person(
+                "Ivan IV The Terrible", 54, 170, 70,
+                new Person("Vasili III of Russia", 28, 170, 60, null));
+        }
+    }
+}
